# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/organizations/{organizationId}/customers": {
      "get": {
        "description": "Requires a verified Auth0 access token for the configured API audience. Organization membership is resolved server-side. A missing membership or foreign resource returns 404; IDs never grant access. Required permissions: customers:read. See docs/security.md for the role matrix and service-level restrictions. DTO validation and documented schemas do not replace authentication or business rules.",
        "operationId": "CustomersController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "default": "active",
              "type": "string",
              "enum": [
                "active",
                "archived",
                "all"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 120,
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "company",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 120,
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customers"
                }
              }
            }
          },
          "400": {
            "description": "Malformed identifiers, DTO validation failure or invalid workflow.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or wrong-audience access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated member lacks permission or violates a role restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Organization membership or tenant-scoped resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth0": []
          }
        ],
        "summary": "Customers: list",
        "tags": [
          "Customers"
        ]
      },
      "post": {
        "description": "Requires a verified Auth0 access token for the configured API audience. Organization membership is resolved server-side. A missing membership or foreign resource returns 404; IDs never grant access. Required permissions: customers:write. See docs/security.md for the role matrix and service-level restrictions. DTO validation and documented schemas do not replace authentication or business rules.",
        "operationId": "CustomersController_create",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "400": {
            "description": "Malformed identifiers, DTO validation failure or invalid workflow.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or wrong-audience access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated member lacks permission or violates a role restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Organization membership or tenant-scoped resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict with existing data or workflow state where applicable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth0": []
          }
        ],
        "summary": "Customers: create",
        "tags": [
          "Customers"
        ]
      }
    },
    "/v1/organizations/{organizationId}/customers/{customerId}": {
      "get": {
        "description": "Requires a verified Auth0 access token for the configured API audience. Organization membership is resolved server-side. A missing membership or foreign resource returns 404; IDs never grant access. Required permissions: customers:read. See docs/security.md for the role matrix and service-level restrictions. DTO validation and documented schemas do not replace authentication or business rules.",
        "operationId": "CustomersController_getOne",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "400": {
            "description": "Malformed identifiers, DTO validation failure or invalid workflow.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or wrong-audience access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated member lacks permission or violates a role restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Organization membership or tenant-scoped resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth0": []
          }
        ],
        "summary": "Customers: get one",
        "tags": [
          "Customers"
        ]
      },
      "patch": {
        "description": "Requires a verified Auth0 access token for the configured API audience. Organization membership is resolved server-side. A missing membership or foreign resource returns 404; IDs never grant access. Required permissions: customers:write. See docs/security.md for the role matrix and service-level restrictions. DTO validation and documented schemas do not replace authentication or business rules.",
        "operationId": "CustomersController_update",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "400": {
            "description": "Malformed identifiers, DTO validation failure or invalid workflow.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or wrong-audience access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated member lacks permission or violates a role restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Organization membership or tenant-scoped resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict with existing data or workflow state where applicable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth0": []
          }
        ],
        "summary": "Customers: update",
        "tags": [
          "Customers"
        ]
      }
    },
    "/v1/organizations/{organizationId}/customers/{customerId}/restore": {
      "post": {
        "description": "Requires a verified Auth0 access token for the configured API audience. Organization membership is resolved server-side. A missing membership or foreign resource returns 404; IDs never grant access. Required permissions: customers:write. See docs/security.md for the role matrix and service-level restrictions. DTO validation and documented schemas do not replace authentication or business rules.",
        "operationId": "CustomersController_restore",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "400": {
            "description": "Malformed identifiers, DTO validation failure or invalid workflow.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or wrong-audience access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated member lacks permission or violates a role restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Organization membership or tenant-scoped resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict with existing data or workflow state where applicable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth0": []
          }
        ],
        "summary": "Customers: restore",
        "tags": [
          "Customers"
        ]
      }
    },
    "/v1/organizations/{organizationId}/customers/{customerId}/archive": {
      "post": {
        "description": "Requires a verified Auth0 access token for the configured API audience. Organization membership is resolved server-side. A missing membership or foreign resource returns 404; IDs never grant access. Required permissions: customers:write. See docs/security.md for the role matrix and service-level restrictions. DTO validation and documented schemas do not replace authentication or business rules.",
        "operationId": "CustomersController_archive",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Tenant-scoped UUID. Ownership and membership are checked server-side.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "400": {
            "description": "Malformed identifiers, DTO validation failure or invalid workflow.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or wrong-audience access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated member lacks permission or violates a role restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Organization membership or tenant-scoped resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict with existing data or workflow state where applicable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth0": []
          }
        ],
        "summary": "Customers: archive",
        "tags": [
          "Customers"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpsDesk API",
    "description": "Tenant-scoped support operations API. Use the Next.js BFF for browser requests; Auth0 bearer tokens belong on the API boundary. See docs/api/README.md for workflows and Socket.IO events. No real credentials or account details are embedded.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:3001",
      "description": "Local API only; configure deployment URLs privately"
    }
  ],
  "components": {
    "securitySchemes": {
      "auth0": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Auth0 access token with the configured API audience and issuer; never an ID token."
      },
      "svix": {
        "type": "apiKey",
        "in": "header",
        "name": "svix-signature",
        "description": "Verified Resend/Svix signature. Also requires svix-id and svix-timestamp; the signature covers the exact raw body."
      }
    },
    "schemas": {
...
      "CreateCustomerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "email": {
            "type": "string",
            "nullable": true,
            "format": "email",
            "maxLength": 254
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "maxLength": 40
          },
          "company": {
            "type": "string",
            "nullable": true,
            "maxLength": 120
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "maxLength": 2000
          }
        },
        "required": [
          "name"
        ],
        "additionalProperties": false
      },
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 120
          },
          "email": {
            "type": "string",
            "nullable": true,
            "format": "email",
            "maxLength": 254
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "maxLength": 40
          },
          "company": {
            "type": "string",
            "nullable": true,
            "maxLength": 120
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "maxLength": 2000
          }
        },
        "additionalProperties": false
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "company": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "archivedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "phone",
          "company",
          "notes",
          "archivedAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "Customers": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string",
                  "nullable": true
                },
                "phone": {
                  "type": "string",
                  "nullable": true
                },
                "company": {
                  "type": "string",
                  "nullable": true
                },
                "notes": {
                  "type": "string",
                  "nullable": true
                },
                "archivedAt": {
                  "type": "string",
                  "format": "date-time",
                  "nullable": true
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                }
              },
              "required": [
                "id",
                "name",
                "email",
                "phone",
                "company",
                "notes",
                "archivedAt",
                "createdAt",
                "updatedAt"
              ]
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "totalPages": {
                "type": "integer"
              },
              "hasNextPage": {
                "type": "boolean"
              },
              "hasPreviousPage": {
                "type": "boolean"
              }
            },
            "required": [
              "page",
              "limit",
              "total",
              "totalPages",
              "hasNextPage",
              "hasPreviousPage"
            ]
          }
        },
        "required": [
          "data",
          "pagination"
        ]
      },
...
    }
  }
}